Worked on Displaying all items in inventory and adding items to a shoppping cart
Removed orderQuantity and stock variables under Item because if we have a list of items to represent our inventory, it wouldn't make sense to have the orderQuantity variable in that list.  And vice versa, if we have a list of items for our shopping cost, it wouldn't make sense for the items to store the stock variable.

Rather, I created two new classes.  ItemStock to be put in an arraylist which will represent our inventory, and ItemOrder which will be put in an arraylist in ShoppinKart.

Added Category as a variable to Item because it wasn't there for some reason.

Implemented a jList which will list all the products we have in itemStockList

Implemented the logic behind adding an item you see in the jList to the currentCustomer's shopping cart.

ShoppingCart changes: Removed the customer attribute from ShoppingCart Class.  Now, ShoppingCart is a private variable inside of CustomerAccount.

Moved and modified the existing shoppingcart methods to the ShoppingCart class. (untested)

Created categoryList, itemStockList, supplierList in main class.